+++ packages/reflex-base/news/6947.bugfix.md
@@ -1,0 +1,1 @@
+Auto-memoized components whose module-level code came from `add_custom_code` no longer collide on a generated memo name. Two otherwise-identical components emitting different custom code shared one memo module, so one of their two code blocks was dropped from the compiled output.
