+++ packages/reflex-base/src/reflex_base/components/memo.py
@@ -3,12 +3,15 @@
 from __future__ import annotations
 
 import dataclasses
+import enum
 import inspect
+import operator
 import sys
 from collections.abc import Callable, Mapping, Sequence
 from copy import copy
 from enum import Enum
 from functools import cache, partial, update_wrapper
+from hashlib import md5
 from types import UnionType
 from typing import (
     Annotated,
@@ -28,7 +31,7 @@ from typing import (
 from reflex_components_core.base.fragment import Fragment
 
 from reflex_base import constants
-from reflex_base.components.component import Component
+from reflex_base.components.component import BaseComponent, Component
 from reflex_base.components.memoize_helpers import (
     MemoizationStrategy,
     get_memoization_strategy,
@@ -1758,6 +1761,305 @@ def _create_component_wrapper(
     return _MemoComponentWrapper(definition)
 
 
+_HASH_BUFFER_FLUSH_SIZE = 1 << 16
+_HASH_MAX_CACHED_STR = 128
+_HASH_MAX_CACHE_ENTRIES = 4096
+
+# Encoded forms of the values that recur across every component hashed during a
+# compile: short strings (dict keys, tags, module paths) and ``ImportVar``
+# instances, which make up the bulk of what a component hash feeds in.
+# ``ImportVar`` is a frozen dataclass whose fields are all ``str``/``bool``/
+# ``None``, so its generated equality means exactly "same encoding" and is safe
+# to key a cache on. Both caches stop admitting new entries at
+# ``_HASH_MAX_CACHE_ENTRIES`` so an app rendering many one-off strings can't
+# grow them without bound; the recurring values get in first and stay.
+_hash_str_encodings: dict[str, bytes] = {}
+_hash_import_var_encodings: dict[ImportVar, bytes] = {}
+_hash_dataclass_layouts: dict[type, tuple[bytes, tuple[tuple[bytes, str], ...]]] = {}
+
+
+def _hash_dataclass_layout(cls: type) -> tuple[bytes, tuple[tuple[bytes, str], ...]]:
+    """Get the cached type tag and pre-encoded field names for a dataclass.
+
+    Args:
+        cls: The dataclass type to describe.
+
+    Returns:
+        The type tag plus field count, and each field's encoded and plain name.
+    """
+    layout = _hash_dataclass_layouts.get(cls)
+    if layout is None:
+        fields = dataclasses.fields(cls)  # pyright: ignore [reportArgumentType]
+        layout = (
+            b"D" + len(fields).to_bytes(8, "little"),
+            tuple((field.name.encode(), field.name) for field in fields),
+        )
+        _hash_dataclass_layouts[cls] = layout
+    return layout
+
+
+def _encode_str_for_hash(value: str) -> bytes:
+    """Encode a string as a type-tagged, length-prefixed payload.
+
+    Args:
+        value: The string to encode.
+
+    Returns:
+        The encoded string.
+    """
+    encoded = value.encode()
+    return b"s" + len(encoded).to_bytes(8, "little") + encoded
+
+
+def _encode_deterministic(value: Any, out: bytearray, hasher: Any | None) -> None:
+    """Append ``value``'s self-delimiting encoding to ``out``.
+
+    Dispatch is on the exact type so the common leaves (strings, bools,
+    containers, imports) skip the ``isinstance`` ladder in
+    :func:`_encode_deterministic_subclass`, which handles everything else.
+    ``out`` is flushed into ``hasher`` at container boundaries once it grows
+    past ``_HASH_BUFFER_FLUSH_SIZE``, so encoding a large subtree never
+    buffers the whole thing.
+
+    Args:
+        value: The value to encode.
+        out: The buffer to append the encoding to.
+        hasher: The hasher ``out`` is flushed into when it grows too large, or
+            ``None`` when ``out`` is a sub-buffer whose full contents the caller
+            needs and so must not be drained mid-encoding.
+    """
+    value_type = type(value)
+    if value_type is str:
+        encoded = _hash_str_encodings.get(value)
+        if encoded is None:
+            encoded = _encode_str_for_hash(value)
+            if (
+                len(value) <= _HASH_MAX_CACHED_STR
+                and len(_hash_str_encodings) < _HASH_MAX_CACHE_ENTRIES
+            ):
+                _hash_str_encodings[value] = encoded
+        out += encoded
+    elif value_type is bool:
+        out += b"T" if value else b"F"
+    elif value_type is dict:
+        out += b"d"
+        out += len(value).to_bytes(8, "little")
+        for k, v in sorted(value.items(), key=operator.itemgetter(0)):
+            _encode_deterministic(k, out, hasher)
+            _encode_deterministic(v, out, hasher)
+        if len(out) > _HASH_BUFFER_FLUSH_SIZE and hasher is not None:
+            hasher.update(out)
+            del out[:]
+    elif value_type is ImportVar:
+        encoded = _hash_import_var_encodings.get(value)
+        if encoded is None:
+            header, fields = _hash_dataclass_layout(ImportVar)
+            buffer = bytearray(header)
+            for encoded_name, name in fields:
+                buffer += encoded_name
+                _encode_deterministic(getattr(value, name), buffer, None)
+            encoded = bytes(buffer)
+            if len(_hash_import_var_encodings) < _HASH_MAX_CACHE_ENTRIES:
+                _hash_import_var_encodings[value] = encoded
+        out += encoded
+    elif value_type is list or value_type is tuple:
+        out += b"l"
+        out += len(value).to_bytes(8, "little")
+        for item in value:
+            _encode_deterministic(item, out, hasher)
+        if len(out) > _HASH_BUFFER_FLUSH_SIZE and hasher is not None:
+            hasher.update(out)
+            del out[:]
+    elif value is None:
+        out += b"N"
+    elif value_type is int or value_type is float:
+        out += b"n"
+        out += str(value).encode()
+    else:
+        _encode_deterministic_subclass(value, out, hasher)
+
+
+def _encode_deterministic_subclass(
+    value: Any, out: bytearray, hasher: Any | None
+) -> None:
+    """Append the encoding of a value whose exact type has no fast path.
+
+    Covers subclasses of the fast-path types — notably ``str``-based enums,
+    which must encode as enums rather than as strings — plus ``Var``,
+    dataclasses, and components.
+
+    Args:
+        value: The value to encode.
+        out: The buffer to append the encoding to.
+        hasher: The hasher ``out`` is flushed into when it grows too large, or
+            ``None`` when ``out`` must not be drained mid-encoding.
+
+    Raises:
+        TypeError: If the value is not hashable.
+    """
+    if isinstance(value, bool):
+        out += b"T" if value else b"F"
+    elif isinstance(value, (int, float, enum.Enum)):
+        out += b"n"
+        out += str(value).encode()
+    elif isinstance(value, str):
+        out += _encode_str_for_hash(value)
+    elif isinstance(value, dict):
+        out += b"d"
+        out += len(value).to_bytes(8, "little")
+        for k, v in sorted(value.items(), key=operator.itemgetter(0)):
+            _encode_deterministic(k, out, hasher)
+            _encode_deterministic(v, out, hasher)
+    elif isinstance(value, (tuple, list)):
+        out += b"l"
+        out += len(value).to_bytes(8, "little")
+        for item in value:
+            _encode_deterministic(item, out, hasher)
+    elif isinstance(value, Var):
+        out += b"v"
+        _encode_deterministic(value._js_expr, out, hasher)
+        _encode_deterministic(value._get_all_var_data(), out, hasher)
+    elif dataclasses.is_dataclass(value):
+        header, fields = _hash_dataclass_layout(
+            value if isinstance(value, type) else type(value)
+        )
+        out += header
+        for encoded_name, name in fields:
+            out += encoded_name
+            _encode_deterministic(getattr(value, name), out, hasher)
+    elif isinstance(value, BaseComponent):
+        out += b"C"
+        _encode_deterministic(value.render(), out, hasher)
+    else:
+        msg = (
+            f"Cannot hash value `{value}` of type `{type(value).__name__}`. "
+            "Only BaseComponent, Var, VarData, dict, str, tuple, and enum.Enum are supported."
+        )
+        raise TypeError(msg)
+    if len(out) > _HASH_BUFFER_FLUSH_SIZE and hasher is not None:
+        hasher.update(out)
+        del out[:]
+
+
+def _update_deterministic_hash(hasher: Any, value: object) -> None:
+    """Feed ``value`` into ``hasher`` using a self-delimiting, type-tagged encoding.
+
+    Each branch writes a distinct type tag plus length-prefixed payload, which
+    keeps the encoding injective without building intermediate strings. The
+    encoding is buffered in a ``bytearray`` and handed to the hasher in large
+    chunks instead of one ``update`` per node, since a single component hash
+    covers tens of thousands of nodes.
+
+    Args:
+        hasher: A ``hashlib`` hasher (must accept ``.update(bytes)``).
+        value: The value to fold into the hasher.
+    """
+    buffer = bytearray()
+    _encode_deterministic(value, buffer, hasher)
+    hasher.update(buffer)
+
+
+def _deterministic_hash(value: object) -> str:
+    """Hash a rendered dictionary.
+
+    Args:
+        value: The dictionary to hash.
+
+    Returns:
+        The hash of the dictionary.
+
+    Raises:
+        TypeError: If the value is not hashable.
+    """
+    hasher = md5(usedforsecurity=False)
+    _update_deterministic_hash(hasher, value)
+    return hasher.hexdigest()
+
+
+def _update_component_artifacts_hash(
+    hasher: Any, component: Component, *, recursive: bool
+) -> None:
+    """Fold a component's compile artifacts into ``hasher``.
+
+    Two components can render identical JSX and still compile to different
+    modules -- the classic case is a differing ``on_mount``, which ``_render``
+    omits but which shows up as a lifecycle hook -- so the imports, hooks,
+    custom code, and app-wrap components have to be part of the hash too.
+
+    Everything is encoded into one shared buffer rather than a hasher update
+    per artifact.
+
+    Args:
+        hasher: A ``hashlib`` hasher to fold the artifacts into.
+        component: The component whose memo body is being hashed.
+        recursive: Whether descendants' artifacts belong to this memo body.
+            False for a passthrough memo, whose descendants render at the call
+            site behind the ``{children}`` hole, so only the component's own
+            artifacts identify the body.
+    """
+    buffer = bytearray()
+    if recursive:
+        _encode_deterministic(component._get_all_imports(), buffer, hasher)
+        _encode_deterministic(component._get_all_hooks_internal(), buffer, hasher)
+        _encode_deterministic(component._get_all_hooks(), buffer, hasher)
+        _encode_deterministic(component._get_all_custom_code(), buffer, hasher)
+        _encode_deterministic(component._get_all_app_wrap_components(), buffer, hasher)
+    else:
+        _encode_deterministic(component._get_imports(), buffer, hasher)
+        _encode_deterministic(component._get_hooks_internal(), buffer, hasher)
+        _encode_deterministic(component._get_hooks(), buffer, hasher)
+        _encode_deterministic(component._get_added_hooks(), buffer, hasher)
+        _encode_deterministic(component._get_custom_code(), buffer, hasher)
+        # ``_get_all_custom_code`` folds in ``add_custom_code`` on the recursive
+        # side; the own-node side has to ask for it explicitly. It used not to,
+        # so two passthrough bodies differing only in ``add_custom_code`` output
+        # collided on a tag.
+        for clz in component._iter_parent_classes_with_method("add_custom_code"):
+            _encode_deterministic(clz.add_custom_code(component), buffer, hasher)
+        _encode_deterministic(component._get_app_wrap_components(), buffer, hasher)
+    hasher.update(buffer)
+
+
+def component_hash(component: Component, *, recursive: bool) -> str:
+    """Get a stable content hash for a component's memo body.
+
+    Args:
+        component: The component being memoized.
+        recursive: Whether the memo body carries the component's whole subtree
+            (a snapshot memo) rather than a ``{children}`` hole.
+
+    Returns:
+        The hex digest content hash.
+    """
+    hasher = md5(usedforsecurity=False)
+    _update_deterministic_hash(hasher, component.render())
+    _update_component_artifacts_hash(hasher, component, recursive=recursive)
+    return hasher.hexdigest()
+
+
+def memo_tag(component: Component) -> str:
+    """Compute a stable tag name for the memo wrapping ``component``.
+
+    The class qualname is encoded directly in the tag prefix so that distinct
+    classes which happen to render identically never collide on a tag. Tag
+    collision would silently share a single cached memo wrapper across classes
+    and drop the later class's class-level metadata (e.g.
+    ``_get_app_wrap_components``, which carries providers like
+    ``UploadFilesProvider`` that must reach the app root).
+
+    Args:
+        component: The component being memoized.
+
+    Returns:
+        The stable tag name.
+    """
+    recursive = get_memoization_strategy(component) is MemoizationStrategy.SNAPSHOT
+    return format.format_state_name(
+        f"{type(component).__qualname__}_{component.tag or 'Comp'}_"
+        f"{component_hash(component, recursive=recursive)}"
+    ).capitalize()
+
+
 def create_passthrough_component_memo(
     component: Component,
     source_module: str | None = None,
@@ -1771,7 +2073,7 @@ def create_passthrough_component_memo(
     through the memo pipeline instead of emitting ad-hoc page-local
     ``React.memo`` declarations.
 
-    The exported memo name is derived from ``component._compute_memo_tag()``
+    The exported memo name is derived from :func:`memo_tag`
     after the ``{children}`` hole has been substituted into the wrapped
     component's children (passthrough mode), so two call-sites differing only
     in their children — whose generated memo bodies are identical — collapse
@@ -1842,7 +2144,7 @@ def create_passthrough_component_memo(
             "normalizes to `rx.Component`."
         )
         raise TypeError(msg)
-    tag = preview._compute_memo_tag()
+    tag = memo_tag(preview)
 
     passthrough.__name__ = format.to_snake_case(tag)
     passthrough.__qualname__ = passthrough.__name__
