+++ tests/units/components/test_component.py
@@ -4,13 +4,8 @@ from dataclasses import dataclass
 from typing import Any, ClassVar, TypedDict
 
 import pytest
-from reflex_base.components.component import (
-    _HASH_MAX_CACHE_ENTRIES,
-    Component,
-    _deterministic_hash,
-    field,
-)
-from reflex_base.constants import EventTriggers, Hooks
+from reflex_base.components.component import Component, field
+from reflex_base.constants import EventTriggers
 from reflex_base.constants.state import FIELD_MARKER
 from reflex_base.event import (
     EventChain,
@@ -2346,107 +2341,3 @@ def test_get_all_hooks_internal_does_not_mutate_hooks_cache():
     assert dict(parent._get_hooks_internal()) == parent_own_hooks
     # And repeated collection yields the same result.
     assert parent._get_all_hooks_internal() == combined
-
-
-def test_deterministic_hash_is_stable():
-    """The same value must hash identically across calls and dict orderings."""
-    value = {"b": [1, "x"], "a": {"k": None}}
-    reordered = {"a": {"k": None}, "b": [1, "x"]}
-
-    assert _deterministic_hash(value) == _deterministic_hash(value)
-    assert _deterministic_hash(value) == _deterministic_hash(reordered)
-
-
-@pytest.mark.parametrize(
-    ("left", "right"),
-    [
-        # Type tags must keep values of different types apart.
-        ("1", 1),
-        (1, True),
-        (0, False),
-        (None, "None"),
-        ({"a": "b"}, [["a", "b"]]),
-        # Length prefixes must keep concatenations apart.
-        (["ab", "c"], ["a", "bc"]),
-        ([[], []], [[[]]]),
-        ({"a": "", "b": ""}, {"ab": ""}),
-        # Nested containers must not flatten into their contents.
-        ([1, [2]], [1, 2]),
-        # str-keyed enums encode as enums, not as their string value.
-        (Hooks.HookPosition.PRE_TRIGGER, Hooks.HookPosition.PRE_TRIGGER.value),
-        # Dataclasses of the same shape but different types stay distinct.
-        (ImportVar(tag="a"), ImportVar(tag="a", alias="a")),
-    ],
-)
-def test_deterministic_hash_distinguishes(left: Any, right: Any):
-    """Distinct values must not collide under the type-tagged encoding."""
-    assert _deterministic_hash(left) != _deterministic_hash(right)
-
-
-def test_deterministic_hash_treats_lists_and_tuples_alike():
-    """Sequences share one type tag, so a list and tuple of equal items match."""
-    assert _deterministic_hash([1, "a"]) == _deterministic_hash((1, "a"))
-
-
-def test_deterministic_hash_import_var_cache_is_by_value():
-    """Equal ``ImportVar`` instances hash the same; unequal ones do not.
-
-    ``ImportVar`` encodings are cached by value, so a stale or over-eager cache
-    entry would show up as two unequal imports hashing alike.
-    """
-    a = ImportVar(tag="useState", is_default=False, install=True)
-    b = ImportVar(tag="useState", is_default=False, install=True)
-    c = ImportVar(tag="useState", is_default=True, install=True)
-
-    assert _deterministic_hash(a) == _deterministic_hash(b)
-    assert _deterministic_hash(a) != _deterministic_hash(c)
-    assert _deterministic_hash({"react": (a, c)}) != _deterministic_hash({
-        "react": (c, a)
-    })
-
-
-def test_deterministic_hash_long_strings():
-    """Strings past the encoding cache's size limit still hash correctly."""
-    long_a = "a" * 10_000
-    long_b = "a" * 9_999 + "b"
-
-    assert _deterministic_hash(long_a) == _deterministic_hash("a" * 10_000)
-    assert _deterministic_hash(long_a) != _deterministic_hash(long_b)
-
-
-def test_deterministic_hash_beyond_string_cache_capacity():
-    """Strings that arrive after the encoding cache fills still hash correctly."""
-    values = [f"cache_capacity_probe_{i}" for i in range(_HASH_MAX_CACHE_ENTRIES + 500)]
-    digests = [_deterministic_hash(value) for value in values]
-
-    assert len(set(digests)) == len(values)
-    assert [_deterministic_hash(value) for value in values] == digests
-
-
-def test_deterministic_hash_flushes_large_payloads():
-    """A payload larger than the buffer flush size hashes deterministically."""
-    payload = {f"key_{i}": "v" * 200 for i in range(2000)}
-
-    assert _deterministic_hash(payload) == _deterministic_hash(dict(payload))
-    mutated = {**payload, "key_0": "w" * 200}
-    assert _deterministic_hash(payload) != _deterministic_hash(mutated)
-
-
-def test_deterministic_hash_components_and_vars():
-    """Components and Vars hash by rendered content, not by identity."""
-    assert _deterministic_hash(Bare.create(contents="a")) == _deterministic_hash(
-        Bare.create(contents="a")
-    )
-    assert _deterministic_hash(Bare.create(contents="a")) != _deterministic_hash(
-        Bare.create(contents="b")
-    )
-    assert _deterministic_hash(Var("a")) == _deterministic_hash(Var("a"))
-    assert _deterministic_hash(Var("a")) != _deterministic_hash(Var("b"))
-    # A Var and the bare string it renders to must not collide.
-    assert _deterministic_hash(Var("a")) != _deterministic_hash("a")
-
-
-def test_deterministic_hash_rejects_unsupported_types():
-    """Values with no encoding raise rather than hashing to a shared digest."""
-    with pytest.raises(TypeError):
-        _deterministic_hash(object())
